Casting to void* and reinterpret cast are evil
Instead of using reinterpret_cast use the safer static_cast
Furthermore in C++ casting to void* is a really unsafe. The
main thing to keep in mind is to use the initial type when
casting back. So we must ensure that
void* t =  static_cast<QObject*>(value)
QObject* t = static_cast<void*>(value)