Fixed broken model instantiated from Qml
This was due by multiple factors. The most important is that we always registered Models as DosQAbstractItemModelWrappers that is a subclass of QAbstractItemModel. This in turn broke the APIs when the user really instantiated a QAbstractTableModel from the binded language because the cast from DosQAsbstratItemModelWrapper to QAbstractTableModel failed.
The solution is to add different correct subclasses during registration